API: change level depth to signed int
Change obj->depth to signed int, and all relevant function manipulating
object depths.

We have negative depth for special levels so we cast in lots of places.
This is becoming even more common with the NUMAnode special level.

Also change hwloc_topology_get_depth() to signed to avoid signedness
comparison warnings when iterating over depths.
That one is less obvious since it's always >0 and even an unsigned internally.

Cache/Bridge/Group specific depth attr are still unsigned,
those are never negative.
It's not clear whether mixing signed depth for object and unsigned
for attribute could cause confusion.

+++ hwloc/topology-synthetic.c
@@ -956,7 +956,7 @@ static int hwloc_topology_export_synthetic_indexes(struct hwloc_topology * topol
 						   hwloc_obj_t obj,
 						   char *buffer, size_t buflen)
 {
-  unsigned depth = obj->depth;
+  int depth = obj->depth;
   unsigned total;
   hwloc_obj_t *level;
   unsigned step = 1;
@@ -968,8 +968,8 @@ static int hwloc_topology_export_synthetic_indexes(struct hwloc_topology * topol
   char *tmp = buffer;
   int res, ret = 0;
 
-  if ((int)depth < 0) {
-    assert((int)depth == HWLOC_TYPE_DEPTH_NUMANODE);
+  if (depth < 0) {
+    assert(depth == HWLOC_TYPE_DEPTH_NUMANODE);
     total = topology->slevels[HWLOC_SLEVEL_NUMANODE].nbobjs;
     level = topology->slevels[HWLOC_SLEVEL_NUMANODE].objs;
   } else {
